+++ src/App.css
@@ -37,12 +37,4 @@
   }
 }
 
-.A {
-background-color:palegreen;
-padding: 1cm
-}
 
-.B {
-  background-color:whitesmoke;
-  border:2px solid teal;
-}
