+++ src/components/AddTask/AddTask.css
@@ -1,0 +1,3 @@
+.С{
+    border: turquoise;
+    }
