+++ src/components/Task/Task.css
@@ -1,0 +1,6 @@
+.B {
+    background-color:whitesmoke;
+    border:2px solid teal;
+    text-align:center;
+    margin: 25px;
+    }
