+++ src/components/ToDoList/ToDoList.css
@@ -1,0 +1,15 @@
+header{
+    text-align: center;
+    background: cyan;
+    padding: 25px;
+}
+.A {
+    background-color:palegreen;
+    text-align:center;
+    }
+
+.D{
+text-align: center;
+background: yellow;
+padding: 25px;
+}
